refactor: minimum supported `Node.js` version is `14.15.0`
