+++ jest.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  testMatch: [
+    "**/__tests__/**/*.(m|c)?[jt]s?(x)",
+    "**/?(*.)+(spec|test).(m|c)?[jt]s?(x)",
+  ],
+  runner: "jest-light-runner",
+};
