+++ package.json
@@ -13,7 +13,7 @@
   },
   "main": "dist/cjs.js",
   "engines": {
-    "node": ">= 12.13.0"
+    "node": ">= 14.15.0"
   },
   "scripts": {
     "start": "npm run build -- -w",
@@ -25,9 +25,9 @@
     "lint:prettier": "prettier \"{**/*,*}.{js,json,md,yml,css,ts}\" --list-different",
     "lint:js": "eslint --cache .",
     "lint": "npm-run-all -l -p \"lint:**\"",
-    "test:only": "cross-env NODE_ENV=test NODE_OPTIONS=--experimental-vm-modules jest --runInBand ./test/loader.test.js",
+    "test:only": "node --loader ts-node/esm ./node_modules/.bin/jest",
     "test:watch": "npm run test:only -- --watch",
-    "test:coverage": "npm run test:only -- --collectCoverageFrom=\"src/**/*.js\" --coverage",
+    "test:coverage": "npm run test:only -- --collectCoverageFrom=\"src/**/*.mjs\" --coverage",
     "pretest": "npm run lint",
     "test": "npm run test:coverage",
     "prepare": "husky install && npm run build",
@@ -49,10 +49,10 @@
     "@babel/core": "^7.14.6",
     "@babel/eslint-parser": "^7.15.8",
     "@babel/preset-env": "^7.14.7",
-    "@commitlint/cli": "^13.1.0",
-    "@commitlint/config-conventional": "^13.1.0",
+    "@commitlint/cli": "^17.0.2",
+    "@commitlint/config-conventional": "^17.0.2",
     "@webpack-contrib/eslint-config-webpack": "^3.0.0",
-    "babel-jest": "^27.0.6",
+    "babel-jest": "^28.1.0",
     "cross-env": "^7.0.3",
     "del": "^6.0.0",
     "del-cli": "^4.0.1",
@@ -60,9 +60,10 @@
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.23.4",
     "html-loader": "^3.0.0",
-    "husky": "^7.0.0",
-    "jest": "^27.0.6",
-    "lint-staged": "^11.0.1",
+    "husky": "^8.0.1",
+    "jest": "^28.1.0",
+    "jest-light-runner": "^0.2.1",
+    "lint-staged": "^13.0.0",
     "memfs": "^3.2.2",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.3.2",
