--- test/cjs.test.js
@@ -1,8 +1,0 @@
-import src from "../src/index.js";
-import cjs from "../src/cjs.js";
-
-describe("CJS", () => {
-  it("should export loader", () => {
-    expect(cjs).toEqual(src);
-  });
-});
