+++ test/helpers/getErrors.mjs
@@ -1,3 +1,3 @@
-import normalizeErrors from "./normalizeErrors.js";
+import normalizeErrors from "./normalizeErrors.mjs";
 
 export default (stats) => normalizeErrors(stats.compilation.errors.sort());
