+++ test/helpers/getExecutedCode.mjs
@@ -1,4 +1,4 @@
-import { execute, readAsset } from "./index.js";
+import { execute, readAsset } from "./index.mjs";
 
 export default (asset, compiler, stats) =>
   execute(readAsset(asset, compiler, stats));
