--- test/helpers/index.js
@@ -1,21 +1,0 @@
-import compile from "./compile.js";
-import execute from "./execute.js";
-import getExecutedCode from "./getExecutedCode.js";
-import getCompiler from "./getCompiler.js";
-import getErrors from "./getErrors.js";
-import getWarnings from "./getWarnings.js";
-import normalizeErrors from "./normalizeErrors.js";
-import readAsset from "./readAsset.js";
-import readsAssets from "./readAssets.js";
-
-export {
-  compile,
-  execute,
-  getExecutedCode,
-  getCompiler,
-  getErrors,
-  getWarnings,
-  normalizeErrors,
-  readAsset,
-  readsAssets,
-};
