+++ test/helpers/readAsset.mjs
